Update README with project structure
